Code/ReportWriter-master/rest_api/models.py
# @OPENSOURCE_HEADER_START@
# MORE Tool 
# Copyright 2016 Carnegie Mellon University.
# All Rights Reserved.
#
# THIS SOFTWARE IS PROVIDED "AS IS," WITH NO WARRANTIES WHATSOEVER.
# CARNEGIE MELLON UNIVERSITY EXPRESSLY DISCLAIMS TO THE FULLEST EXTENT
# PERMITTEDBY LAW ALL EXPRESS, IMPLIED, AND STATUTORY WARRANTIES,
# INCLUDING, WITHOUT LIMITATION, THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE, AND NON-INFRINGEMENT OF PROPRIETARY
# RIGHTS.
#
# Released under a modified BSD license, please see license.txt for full
# terms. DM-0003473
# @OPENSOURCE_HEADER_END@
from django.db import models
from django.core.validators import URLValidator
from solo.models import SingletonModel

class RESTConfiguration(SingletonModel):
    url = models.CharField(validators=[URLValidator()], max_length=255)
    token = models.CharField(max_length=255)

    def __unicode__(self):
        return u"REST Configuration"

    class Meta:
        verbose_name = "REST Configuration"

