Code/ReportWriter-master/requirements.txt
Django==1.8.2
django-admin-bootstrapped==2.5.0
django-autocomplete-light==2.2.0rc4
psycopg2==2.6
pytz==2015.4
django-widget-tweaks==1.3
django-recaptcha==1.0.4
mock==1.0.1
django-allauth==0.20.0
django-fluent-comments==1.0.2
django-contrib-comments==1.6.1
-e git+https://github.com/maraujop/django-crispy-forms.git@dev#egg=django-crispy-forms
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.3.0
static3==0.6.1
wsgiref==0.1.2
django-solo==1.1.0
selenium==2.46.0
django-postgrespool==0.3.0
